Move auth from config to proxy
