+++ commands/describe.go
@@ -11,7 +11,6 @@ import (
 	"strings"
 	"text/tabwriter"
 
-	"github.com/openfaas/faas-cli/config"
 	"github.com/openfaas/faas-cli/proxy"
 	"github.com/openfaas/faas-cli/schema"
 	"github.com/openfaas/faas-cli/stack"
@@ -65,7 +64,7 @@ func runDescribe(cmd *cobra.Command, args []string) error {
 		}
 	}
 	gatewayAddress := getGatewayURL(gateway, defaultGateway, yamlGateway, os.Getenv(openFaaSURLEnvironment))
-	cliAuth := config.NewCLIAuth(token, gatewayAddress)
+	cliAuth := proxy.NewCLIAuth(token, gatewayAddress)
 	transport := GetDefaultCLITransport(tlsInsecure, &commandTimeout)
 	cliClient := proxy.NewClient(cliAuth, gatewayAddress, transport, &commandTimeout)
 	ctx := context.Background()
